Create a generic function to read jinja2 templates
Create a generic function to read template files, which are also treated as
jinja2 templates whether they have any custom variables or not. The goal is
for this function to be used across the codebase. At the moment it is
located in emailconfig.py, but I'd like to move it to another config file
at some point along with the rest of the template options.