Refactor and rename login_submission_legacy_convert
Additionally moves some checks that were in
`LoginRestServlet._do_other_login` to
`convert_client_dict_legacy_fields_to_identifier` as they would be
applicable to any other callers of this function.

Additionally, there is a functional change here, in that we now return
M_MISSING_PARAM's instead of M_UNKNOWN errcodes for when `identifier` or
`type` parameters are missing.