Address review comments
Generalise public rooms check, allow blocking removal from public rooms dir as well.

Remove the check on alias association, that is unrelated.

Note that we now execute the same behaviour as when publishing the room is
blocked by the homeserver config. That is, we now 403 the /createRoom call
instead of just letting the room be created without entering it into the public
rooms directory.

Finally, only call check_visibility_can_be_modified if it exists.