Reduce duplication in event ordering lookup methods
`get_topological_token_for_event` and `get_event_ordering` were doing almost
the same thing: the difference was only in the format of the result (`tuple` vs
`RoomStreamOrdering`).

This commit combines the two. `get_event_ordering` had a cache which may have
been tuned so we stick with that name, but update it to return a
`RoomStreamOrdering` for better type safety. We also put it in
`StreamWorkerStore` rather than `EventsWorkerStore` since that's where all the
other "ordering" methods are.

So then we have to update all the places `get_topological_token_for_event` was
called, to call `get_event_ordering` instead, and we have to update all the
places `get_event_ordering` was called (just `is_event_after`) to handle
`RoomStreamToken`s intstead of tuples.

`is_event_after` was only called in once place and we may as well inline it.