Optimise createRoom with multiple invites
By not dropping the membership lock between invites, we can stop joins from
grabbing the lock when we're half-done and slowing the whole thing down.

(Is this a DoS vector? If someone invites thousands of people to a room, they
will receive the invite, but won't be able to accept it until the invites have
stopped churning? Though we should guard against abusive createRooms anyway.)