Generalise _maybe_store_room_on_invite
There's a handy function called maybe_store_room_on_invite which allows us to
create an entry in the rooms table for a room and its version for which we
aren't joined to yet, but we can reference when ingesting events about.

This is currently used for invites where we receive some stripped state about
the room and pass it down via /sync to the client, without us being in the
room yet.

There is a similar requirement for knocking, where we will eventually do the
same thing, and need an entry in the rooms table as well. Thus, reusing this
function works, however its name needs to be generalised a bit.