Generalise _locally_reject_invite
_locally_reject_invite generates an out-of-band membership event which
can be passed to clients but not other homeservers.

This is used when we fail to reject an invite over federation, so
instead just generate a leave event locally and send it down /sync,
allowing clients to reject invites even if we can't reach the remote
homeserver.

A similar flow needs to be put in place for rescinding knocks. If we're
unable to contact any remote server from the room we've tried to knock
on, we'd still like to generate and store the leave event locally. Hence
the need to reuse, and thus generalise, this method.