Optimise retrieving previous invite or knock event during membership update.
It was pointed out that the method in
RoomMemberHandler.update_membership_locked for retrieving the previous invite
or knock event was overly complex. We already knew the ID of the room, yet got
a list of all rooms that the user has been invited to/knocked on from the
database. We then filtered that list, and extracted some metadata from it that
allowed us to get the invite/knock event ID.

This was quite ineffecient, and I've updated both methods to simply get the
current state of the room given the room ID, extract the latest m.room.member
event for the user, and check that it has invite/knock membership.

There's still some inefficiency between pulling out the knock event, passing
the event ID to remote_rescind_knock, to which remote_rescind_knock then
pulls the associated event out of the database again - however I think this
is necessary given that we may need to call remote_rescind_knock over
replication, where passing an event ID is easier than a serialised event.
It also more closely resembles remote_reject_invite's function signature.