Add tests
There didn't seem to be any tests for unread push count yet, so there was
a fair amount of code to write. It didn't help that the push code is a
little arcane.

There's two new tests. One that tests the existing behaviour that groups
unread notification count by room. And another for the new behaviour
which tests unread count by message.

To test unread counts, we needed to have a read receipt exist in the
room. Note that a push notification will be sent out *both* when a
message is received, and when a read receipt is sent (so that the
badge on user's phone apps will update after they've read some
rooms).

We also need to advance the reactor after every action that may cause
a notification.