Fix FederationKnockingTestCase so that it no longer overrides event_auth.check for all tests
We needed to disable event_auth checking for this test. However, the method in
which we were doing so was disabling event_auth.check across all tests. Thus
after this test ran, other unit tests failed in weird and wonderful ways.

Take a different approach which acts on the homeserver created for this test,
rather than mocking a file directly.